Added support for auth based plugins
